// types.ts
export type DiscussionCategory = 'Questions' | 'Discussions' | 'Help' | 'Tutorials' | 'General';
export type DiscussionDifficulty = 'Beginner' | 'Intermediate' | 'Advanced';
export type DiscussionStatus = 'Unsolved' | 'Solved';
export type DiscussionLanguage = 'JavaScript' | 'Python' | 'Java' | 'C++' | string; // etc.

export interface DiscussionThread {
  id: string;
  title: string;
  content: string;
  authorId: string;
  authorName: string;
  authorAvatar: string;
  category: DiscussionCategory;
  language: DiscussionLanguage;
  difficulty: DiscussionDifficulty;
  status: DiscussionStatus;
  timestamp: any; // Firestore Timestamp
  replyCount: number;
  tags?: string[];
  upvotes: number;
  downvotes: number;
  upvotedBy: string[];
  downvotedBy: string[];
  repliedBy: string[];
}

export interface DiscussionReply {
   id: string;
   threadId: string;
   authorId: string;
   authorName: string;
   authorAvatar: string;
  content: string;
  timestamp: any; // Firestore Timestamp
  upvotes: number;
  downvotes: number;
  upvotedBy: string[];
  downvotedBy: string[];
}